+++ Changelog.md
@@ -1,3 +1,6 @@
+# Version 1.3.11
+- display additional info for CV29
+
 # Version 1.3.10
 - Correct Functions load automatically when loco selected from the list
 - Functions return to default when loco dropped